update ui upload nodes
